Moved redundant code between gauge_force_reference and hisq_force_reference into a new force_utils file, which itself is not permanent
